refactor: complete SRP refactoring with 3-module architecture
- Phase 6: Merge event_handler.rs into window_manager.rs for simplicity
- Phase 7: Update WindowManager to use WindowState + WindowRenderer architecture
- Phase 8: Remove obsolete window_operations.rs module

Key improvements:
- WindowState: Pure state management (focus, BSP tree, fullscreen, config)
- WindowRenderer: X11 operations and visual rendering
- WindowManager: Event coordination and delegation
- Fixed window positioning bug where windows stacked at top-left
- All 66 tests pass, interactive testing confirms proper BSP layout

Architecture benefits:
- Single Responsibility Principle: Each module has focused purpose
- Dependency injection: WindowRenderer operates on WindowState
- Maintainability: Clear separation of concerns (data vs rendering vs coordination)
- Reduced complexity: 579 lines removed while preserving all functionality
