+++ src/main.rs
@@ -8,10 +8,8 @@ use tracing::info;
 // Module declarations
 mod bsp;
 mod config;
-mod event_handler;
 mod keyboard;
 mod window_manager;
-mod window_operations;
 mod window_renderer;
 mod window_state;
